Add files via upload
You will need to install the following libraries in Arduino for this to work: AdaFruit BusIO version 1.16.1, Adafruit PM25 AQI and its dependencies. When you get it running, it takes samples of the number of particles present. It then calculates equivalent particle diameter and the number of particles with different diameter per unit volume.